+++ ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: jlebard <jlebard@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/09 17:42:00 by marvin            #+#    #+#             */
-/*   Updated: 2023/11/17 11:42:17 by jlebard          ###   ########.fr       */
+/*   Updated: 2023/11/20 11:32:33 by jlebard          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcat(char *dest, const char	*src, size_t	n)
+size_t	ft_strlcat(char *dest, const char	*src, size_t n)
 {
 	size_t	i;
 	size_t	j;
@@ -21,7 +21,7 @@ size_t	ft_strlcat(char *dest, const char	*src, size_t	n)
 	j = 0;
 	while (dest[i] != '\0')
 		i++;
-	while (src[j] != '\0' && j < (int)n)
+	while (src[j] != '\0' && j < n)
 	{
 		dest[i + j] = src[j];
 		j++;
